+++ CHANGELOG.md
@@ -9,6 +9,19 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 None
 
+<a name="4.2.0"></a>
+
+## [4.2.0] - 2025-09-23
+
+### Added
+
+- Add models for ExportResult and ExportRequest
+
+### Changed
+
+- CallResult.callStatus -> CallResult.status
+- CreateConference now includes parentCallId
+
 <a name="4.1.2"></a>
 
 ## [4.1.2] - 2025-06-30